requirements.txt
langchain==0.3.8
langchain_groq==0.2.1
numpy==2.1.3
pandas==1.5.3
pymongo==4.6.0
python-dotenv==1.0.1
streamlit==1.22.0
